link specifications to document status page
