+++ conformance-search/src/components/SpecsCovered.tsx
@@ -1,5 +1,7 @@
 import clsx from "clsx";
 import { Coverage } from "@/types/json";
+import { addOrdinalSuffix } from "@/utils";
+import { FaExternalLinkAlt } from "react-icons/fa";
 import { spec_info as specs } from "../../data/coverage.json";
 
 export default function SpecsCovered({
@@ -14,24 +16,31 @@ export default function SpecsCovered({
             <div className={clsx("border-b-2 px-3 py-2", !compact && "text-lg")}>
                 <b>{specs.length}</b> specifications covered
             </div>
-            <div className="max-h-[250px] max-w-xl divide-y-1 overflow-y-auto">
+            <div className="flex max-h-[250px] max-w-xl cursor-pointer flex-col divide-y-1 overflow-y-auto">
                 {specs.map((info: Coverage["spec_info"][0]) => (
-                    <div
+                    <a
                         key={info.ISO}
                         className={clsx(
-                            "truncate px-3 py-2 font-light hover:bg-gray-100",
+                            "group flex items-center justify-between gap-2 truncate px-3 py-2 font-light hover:bg-gray-100",
                             compact && "!py-1 text-sm font-normal"
                         )}
+                        href={info.link}
+                        rel="noreferrer"
+                        target="_blank"
                     >
-                        ISO/IEC {info.ISO} (<span className="font-bold">{info.name}</span>): ver
-                        {info.version}
-                        {info.amendmends &&
-                            info.amendmends.length > 0 &&
-                            `, ${info.amendmends.join(", ")}`}
-                        {info.corrigenda &&
-                            info.corrigenda.length > 0 &&
-                            `, ${info.corrigenda.join(", ")}`}
-                    </div>
+                        <span>
+                            ISO/IEC {info.ISO}:{info.date} (
+                            <span className="font-bold">{info.name}</span>):{" "}
+                            {addOrdinalSuffix(info.version)} Edition{" "}
+                            {info.amendmends &&
+                                info.amendmends.length > 0 &&
+                                `, ${info.amendmends.join(", ")}`}
+                            {info.corrigenda &&
+                                info.corrigenda.length > 0 &&
+                                `, ${info.corrigenda.join(", ")}`}
+                        </span>
+                        <FaExternalLinkAlt className="text-xs opacity-0 transition-opacity duration-200 group-hover:opacity-100" />
+                    </a>
                 ))}
             </div>
         </div>
