+++ conformance-search/src/lib/database.ts
@@ -8,7 +8,76 @@ import {
     SearchResultRefined
 } from "@/types";
 import { Dictionary, FileMetadata, Files, Hierarchy } from "@/types/json";
-import { collectMatchingBranches, crawlBranches } from "./utils";
+
+/**
+ * This will go through the given hierarchies and collect all the branches that match the given filter.
+ * We terminate each branch at the first match. So that next filter can start from that branch.
+ * @param filter - the filter to match
+ * @param hierarchies - the hierarchies to search
+ * @param matches - the matches found so far
+ */
+function collectMatchingBranches(
+    filter: Filter,
+    hierarchies: Hierarchy[],
+    matches: Hierarchy[] = []
+) {
+    // Basically, we will start from the root, and go down the tree. If we encounter a node that matches the filter, we will keep the descedants of that node in a list (preserving the branch) and continue to other branches.
+    const { type, value } = filter;
+
+    // * go through hierarchies
+    hierarchies.forEach((hierarchy) => {
+        Object.keys(hierarchy).forEach((key) => {
+            const [keyFourcc, keyType] = key.split(":");
+
+            let found = false;
+            if (type === "fourcc") {
+                if (keyFourcc === value) {
+                    matches.push(hierarchy[key]);
+                    found = true;
+                }
+            } else if (type === "type") {
+                if (keyType === value) {
+                    matches.push(hierarchy[key]);
+                    found = true;
+                }
+            }
+
+            if (Object.keys(hierarchy[key]).length > 0 && !found) {
+                collectMatchingBranches(filter, [hierarchy[key]], matches);
+            }
+        });
+    });
+
+    return matches;
+}
+
+/**
+ * This goes through the given hierarchies and collects all unique the fourccs, types and tuples.
+ * @param branches - the hierarchies to crawl
+ */
+function crawlBranches(
+    branches: Hierarchy[],
+    fourccs: string[] = [],
+    types: string[] = [],
+    tuples: string[] = []
+) {
+    branches.forEach((branch) => {
+        Object.keys(branch).forEach((key) => {
+            const box = branch[key];
+            const [boxFourcc, boxType] = key.split(":");
+
+            if (!fourccs.includes(boxFourcc)) fourccs.push(boxFourcc);
+            if (!types.includes(boxType)) types.push(boxType);
+            if (!tuples.includes(key)) tuples.push(key);
+
+            if (Object.keys(box).length > 0) {
+                crawlBranches([box], fourccs, types, tuples);
+            }
+        });
+    });
+
+    return { fourccs, types, tuples };
+}
 
 class Database {
     static instance: Database;
