+++ data/standard_features/14496-12/spec_info.json
@@ -2,5 +2,6 @@
   "ISO": "14496-12",
   "name": "ISOBMFF",
   "version": 7,
+  "date": 2022,
   "mp4ra_reference": ["ISO"]
 }
