+++ data/standard_features/14496-14/spec_info.json
@@ -2,5 +2,6 @@
   "ISO": "14496-14",
   "name": "MP4",
   "version": -1,
+  "date": -1,
   "mp4ra_reference": ["MP4v1", "MP4v2"]
 }
