+++ data/standard_features/14496-30/spec_info.json
@@ -2,5 +2,6 @@
   "ISO": "14496-30",
   "name": "Timed Text",
   "version": -1,
+  "date": -1,
   "mp4ra_reference": ["ISO-Text"]
 }
