+++ data/standard_features/23008-12/spec_info.json
@@ -2,5 +2,6 @@
   "ISO": "23008-12",
   "name": "HEIF",
   "version": -1,
+  "date": -1,
   "mp4ra_reference": ["HEIF"]
 }
